better JS sdk specs
